Add grid generator
The generator takes a seed and size and generates a stable random grid
of letters based on the frequency of letter usage (e.g. the more
frequently a letter is used, the higher chance it has to get picked
randomly). The seed and size can be pulled from URL params (id and size,
respectively), and will default to generating a 5x5 grid using a unique
seed for the current day (according to the users current time).