Generate UI from state
The grid UI is now generated from state, instead of on the fly. This
decopules the UI from the state required to generate the UI.

The styles for the path have been updated and a header and footer have
been added as well. Still need to implement scoring, reset, undo and
swapping of letters.