Don't allow path to cross when making a selection
Also simplifies and fixes various bugs with swapping behavior. Swap is
now implemented through double tap, rather than press and hold. Tap once
to select, tap again to mark for a swap, tap again to de-select. Once a
cell is marked for swap, a different, valid cell can be tapped on to
perform a swap.