Improve path validation logic. Fixes #1
During selection, a visual indicator will now be displayed for how the
existing path will link to the selection. If the selection is not valid,
there will be no link between it and the existing path. The linking UX
is also useful for determining how the selection will link to the path
in the case where the word is drawn in reverse, or the word is spelled
in reverse.