Consolidate logic for selection anchoring
Change preference from anchoring at end to anchoring at beginning of
selection. This means if you begin by selecting a cell that neighbors
the existing path, the anchor will be the first cell you select.
Otherwise, if you begin the selection with a non-neighboring cell, the
last cell you select will be the anchoring cell.