Fix false positive isCrossing
Since the code has been updated to check each cell as target and source,
it's not longer necessary to test multiple directions on each, only the
direction they would link together.

Also removed the styling that hides options on a share link such as
delete and settings. Don't see a reason to hide them.