Refactor the character generation algorithm
Previously the algorithm would pick a random character based on the
frequency of character usage in the english dictionary. Instead, it now
picks random words from the dictionary until it gets gets exactly the
number of characters it needs to generate the grid. It then randomly
picks characters from that list.

In theory this guarantees that every grid can be 100% completed,
however, in order to spell the words that were used to seed the grid
you'd presumably have to do a lot of character swapping, so in effect
the gameplay should feel very similar to the previous algorithm.