Properly load and unload dictionaries
Dictionaries are now loaded and unloaded based on a combination of user
preference and the words that were validated in the grid. This means
that a user who enabled profanity and used profane words will have that
dictionary listed in their share text and that dictionary will be
enabled for that share URL. Locally, enabling and disabling of the
profanity dictionary will enable and disable the validation of those
words in the grid.

Note that if the profanity dictionary is disabled after profane words
are included, future profane words will not be validated, but the
existing words will cause the share text to indicate profane words were
used, and the dictionary will be loaded for that share URL, until the
profane words are removed from the grid.