Update the algorithm used to generate the grid. Fixes #4 (#16)
This is a rather large changeset that contains several features. The
biggest change is the introduction of a new algorithm for grid
generation that ensures every grid is able to be filled out 100%.
Additionally, there are now two puzzle modes. The default mode is now
much simpler/more casual. It involves trying to figure out a single path
through the grid using the words used to build that path. It tracks only
how many of those words the user has guessed, and accepts only the words
used to generate the grid. There is no undo or reset and no swapping
letters or scoring points.

The other mode is challenge mode, which is essentially how the game used
to work, except that it is now built on top of the new algorithm.
Challenge mode is now also always guaranteed to be able to be filled out
100%, although in order to do that letters may need to be swapped. If
any secret words are spelled in this mode, they will be revealed on the
grid and any swapped letters contained in them will not be penalized
when calculating points.