Revise how revealed letters work
All letters in the next secret word contained in the spelled word are
now revealed. Previously, it would only reveal letters in the secret
word that overlapped consecutively with letters in the spelled word.

For example, if the secret word was SAG and you spelled a word that
included all of those letters but not at the start or end of the word
you spelled, they would have previously not been revealed, but now they
will be.